requirements.txt
# ============================================================
# Financial Risk Intelligence Copilot — Python Dependencies
# ============================================================

# --- Core ---
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.10.4

# --- ML Engine ---
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
xgboost==2.1.3
shap==0.46.0
joblib==1.4.2
mlflow==2.19.0

# --- RAG Engine ---
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2
PyPDF2==3.0.1

# --- LLM Layer ---
openai==1.58.1

# --- API ---
fastapi==0.115.6
uvicorn[standard]==0.34.0

# --- Testing ---
pytest==8.3.4
httpx==0.28.1

# --- Monitoring ---
prometheus-client==0.21.1
